lab6/requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.3
contourpy==1.3.2
cycler==0.12.1
fonttools==4.60.1
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
termcolor==3.1.0
tzdata==2025.2
urllib3==2.5.0